requirements.txt
ansicolor
requests
BeautifulSoup4
regex
htmlmin
compressinja
lxml
retry
selenium
IMAPClient==0.12
toolz
git+https://github.com/ulikoehler/cffi_re2.git
simplejson
tqdm
xlsxwriter
bottle
googletrans
polib
openpyxl
google-cloud-translate
